FIX: changes after code review
