+++ app/assets/javascripts/discourse/app/lib/sidebar/base-custom-sidebar-section.js
