+++ app/assets/javascripts/discourse/app/lib/sidebar/custom-sections.js
@@ -1,8 +1,10 @@
-import BaseSectionHeader from "discourse/lib/sidebar/base-section-header";
-import BaseSectionLink from "discourse/lib/sidebar/base-section-link";
+import BaseCustomSidebarSection from "discourse/lib/sidebar/base-custom-sidebar-section";
+import BaseCustomSidebarSectionLink from "discourse/lib/sidebar/base-custom-sidebar-section-link";
 
 export const customSections = [];
 
 export function addSidebarSection(func) {
-  customSections.push(func.call(this, BaseSectionHeader, BaseSectionLink));
+  customSections.push(
+    func.call(this, BaseCustomSidebarSection, BaseCustomSidebarSectionLink)
+  );
 }
