+++ app/assets/javascripts/discourse/app/templates/components/sidebar/section.hbs
@@ -18,7 +18,7 @@
           (array @headerModel)
           (if @headerModels @headerModels (array))
         }}
-        @class="sidebar-section-header-link"
+        class="sidebar-section-header-link"
         title={{@headerLinkTitle}}
       >
